Convert ClassEntity to ScopedClassEntity
